% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries_naming_conventions}
\alias{countries_naming_conventions}
\title{International naming conventions of countries}
\format{
A data frame with 249 rows and 8 columns:
\describe{
\item{short_name}{Short name of the country}
\item{name}{Name of the country}
\item{alpha.2}{2-letter abbreviation}
\item{alpha.3}{3-letter abbreviation}
\item{region}{Region}
\item{sub.region}{Sub region}
\item{official_name}{Official name}
\item{common_name}{Common name}
}
}
\usage{
countries_naming_conventions
}
\description{
The list of official naming of countries, their regions and abbreviations
}
\keyword{datasets}
